rename TimelineScrollList to LabelTab
Planning to refactor this component into two components: LabelTab (parent) and TimelineScrollList (child). Then the Details page can be a child of LabelTab.
